Update MrzParser.java
The failure its that its doing an OR instead of an AND
if (digit != checkDigit || (checkDigit != FILLER && checkDigit != '0' && digit == '0') ) {
throw new MrzParseException("Check digit verification failed for " + fieldName + ": expected " + digit + " but got " + checkDigit, mrz, new MrzRange(col, col + 1, row), format);
}
According to Wikipedia https://en.wikipedia.org/wiki/Machine-readable_passport
"Check digit over digits 29–42 (may be < if all characters are <)" which if you have an OR this condition its going to fall into thrown the exception.

The other bug is regarding to the parseName(MrzRange range) function. In some documents like the ID from Kenya there is no surnames in the line for name and surname so an out of range exception is thrown. The function has been modified to avoid this incident.